test: migrate SearchForm tests in source
